feat: modify useSignUpHooks
